Experiment with checked index and vec
We can build a type sace wrapper for a Vec and its corresponding index
(usize), so that the user would not accidentally use it for another
Vec. It is conceptually a type of a pointer with vector index.

It is theoretically possible, but not sure if it's worth the effort.