Allow array literals (cannot contain initializers including vars)
This is very interesting. We have AST interpreter, which I thought
we are going to ditch after the bytecode interpreter is implemented,
but it is useful in compile time evaluation.

Because compile time constant has slightly different semantics,
we don't have to use exactly the same interpreter anyway.
If we build the bytecode for the compile time constant and run it,
the compile time would be wasteful because we are going to evaluate
it only once.

And it is surprising how easy it is to integrate compile time
constant evaluation if we had AST interpreter already.
In particular,

    var a = [1 + 2];

will be compiled to

    var a = [3];

in the bytecode.