The first attempt to implement named function arguments
Currently only works with tree walker interpreter.
We need to think a little bit to make it work with bytecode, since we
wouldn't like to resolve argument names at runtime.
Since we already have names in compile time, we should be able to
resolve it at compile time and reduce it to argument ordering.