Wrap FnDecl's stmts into an Rc
Since it is shared among AST and each FnDecl, the whole AST subtree
in the function declaration would be cloned unless we wrap it in an Rc.
This became an issue since we replaced &'ast with a clone in order to
remove circular lifetimes.

One possible drawback would be that Rc cannot be Sync so you cannot
access the AST from multiple threads.