First attempt to make structured control flow like Wasm
In WebAssembly, jump instruction does not contain explicit address,
which makes the compiler much easier to implement, because it does not
have to "fixup" the jump address if it is jumping forward.
The work of calculating the destination address is responsibility of the
runtime virtual machine. This design is a necessity rather than a
preference since Wasm uses variable length encoding for integers,
calculating offset later would shift a whole chunk of instructions.

I like this design so much so that I steal the idea of structural
control flow. Now it looks like a hybrid of higher level language and
assembly.